Remove barcode detection and add webcam streaming
