+++ requirements.txt
@@ -1,5 +1,3 @@
 Flask
 opencv-python-headless
 numpy
-pyzbar
-python-dotenv
