+++ templates/base.html
@@ -2,7 +2,7 @@
 <title>WareEye</title>
 <link href="https://cdn.jsdelivr.net/npm/tailwindcss@2.2.19/dist/tailwind.min.css" rel="stylesheet">
 <nav class="bg-gray-800 p-4 text-white">
-  <a href="{{ url_for('detector.index') }}" class="mr-4">Home</a>
+  <a href="{{ url_for('webcam.index') }}" class="mr-4">Home</a>
   <a href="{{ url_for('cameras.list_cameras') }}" class="mr-4">Camera List</a>
   <a href="{{ url_for('cameras.new_camera_form') }}" class="mr-4">Add Camera</a>
 </nav>
